requirements.txt
python-dotenv
mysql-connector-python
openai
dateparser